+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - added Copy link context menu action to links in messages
 - correctly rendering text on the same line as number in ordered lists in message markdown
 - added resumeTask method to TaskContext for extension API
+- added task-sidebar-item-badges UI component placement for extensions
 
 ## [0.75.0]